Move `OsStringExt` and `OsStrExt` to `std::os`
